requirements.txt
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
fiona==1.10.0
geopandas==0.14.4
idna==3.8
networkx==3.3
numpy==1.26.4
osmium==3.7.0
packaging==24.1
pandas==2.2.2
pygame==2.6.0
PyOpenGL==3.1.7
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
